yoink out card rendering and visioner
